added new post method, update price to be double instead of long

+++ src/main/java/com/dailycodebuffer/ProductService/model/ProductRequest.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class ProductRequest {
    private String name;
-   private long price;
+   private double price;
    private long quantity;
 }
